third_party/WebKit/LayoutTests/inspector/tracing/timeline-model.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script src="../tracing-test.js"></script>
<script>
function test()
{
    var sessionId = "4.20";
    var mainThread = 1;
    var pid = 100;

    var commonMetadata = [
        {
            "args": { "sessionId": sessionId },
            "cat": "disabled-by-default-devtools.timeline",
            "name": "TracingStartedInPage",
            "ph": "I",
            "pid": pid,
            "tid": mainThread,
            "ts": 100,
        },
        {
            "args": { "layerTreeId": 17 },
            "cat": "disabled-by-default-devtools.timeline",
            "name": "SetLayerTreeId",
            "ph": "I",
            "pid": pid,
            "tid": mainThread,
            "ts": 101,
        }
    ];

    var traceEvents = [
        {"name": "Program", "ts": 1000000, "dur": 9999, "ph": "X", args: {},  "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline" },
        {"name": "FunctionCall", "ts": 1000001, "dur": 9998, "ph": "X", args: {}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline" },
        {"name": "ResourceSendRequest", "ts": 1000002, "ph": "I", args: {"data":{"requestId": 1, "url": "http://example.com", "requestMethod": "GET"}}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline" },
        {"name": "RecalculateStyles", "ts": 1001003, "dur": 997, "ph": "X", args: {}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"},
        {"name": "Layout", "ts": 1002001, "ph": "B", args: {beginData: {}}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"},
        {"name": "Layout", "ts": 1003000, "ph": "E", args: {endData: {}}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"},

        {"name": "Program", "ts": 2000000, "ph": "B", args: {},  "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline" },
        {"name": "FunctionCall", "ts": 2000001, "ph": "B", args: {}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline" },
        {"name": "ResourceSendRequest", "ts": 2000002, "ph": "I", args: {"data":{"requestId": 1, "url": "http://example.com", "requestMethod": "GET"}}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline" },
        {"name": "RecalculateStyles", "ts": 2001003, "ph": "B", args: {}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"},
        {"name": "RecalculateStyles", "ts": 2002000, "ph": "E", args: {}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"},
        {"name": "Layout", "ts": 2002101, "ph": "B", args: {beginData: {}}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"},
        {"name": "Layout", "ts": 2003001, "ph": "E", args: {endData: {}}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"},
        {"name": "FunctionCall", "ts": 2009999, "ph": "E", args: {}, "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline" },
        {"name": "Program", "ts": 2009999, "ph": "E", args: {},  "tid": mainThread, "pid": 100, "cat":"disabled-by-default.devtools.timeline"}
    ];

    var recordsFromEvent = [];
    function onRecordAdded(event)
    {
        recordsFromEvent.push(event.data);
    }
    var tracingTimelineModel = InspectorTest.tracingTimelineModel;
    tracingTimelineModel.addEventListener(WebInspector.TimelineModel.Events.RecordAdded, onRecordAdded);
    tracingTimelineModel.setEventsForTest(sessionId, commonMetadata.concat(traceEvents));
    var records = tracingTimelineModel.records();
    InspectorTest.assertEquals(records.length, recordsFromEvent.length);
    for (var i = 0; i < records.length; ++i) {
        InspectorTest.assertEquals(records[i], recordsFromEvent[i]);
        InspectorTest.dumpTimelineModelRecord(records[i]);
    }
    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
Test trace-specific implementation of timeline model
</p>

</body>
</html>
