third_party/WebKit/LayoutTests/inspector/tracing/timeline-model-expected.txt

Test trace-specific implementation of timeline model

Program: 
----> FunctionCall: 
--------> ResourceSendRequest: example.com
--------> RecalculateStyles: 
--------> Layout: 
Program: 
----> FunctionCall: 
--------> ResourceSendRequest: example.com
--------> RecalculateStyles: 
--------> Layout: 
--------> ParseHTML: example.com [778…]
